+++ fladle-plugin/src/main/java/com/osacky/flank/gradle/FladlePluginDelegate.kt
@@ -219,7 +219,7 @@ class FladlePluginDelegate {
     get() = configurations.getByName(FLADLE_CONFIG)
 
   companion object {
-    val GRADLE_MIN_VERSION: GradleVersion = GradleVersion.version("6.5")
+    val GRADLE_MIN_VERSION: GradleVersion = GradleVersion.version("8.0")
     const val TASK_GROUP = "fladle"
     const val FLADLE_CONFIG = "fladle"